Fix unit tests which were failing on other platforms
BuildCommandLine() was sporadically throwing errors during tests because it was being run more than once... which triggered the validation for unique CLI options. Caching the result fixes those tests.

Upgraded the line endings normalization function.

Fixed a bug where the unix root path was accidently set to an empty string, which does not exist and caused globs to fail.