Fixes #116
Takes the work from @noahmacaskill and rebuilds it - It was a little bit tricky to work out where to put the validation for the output and clobber options.

The validation is in options, and the creation of the text writer is still in the OutputSink.

The OutputSink got changed into a class so we could properly inject dependencies into it.