Adds support for reading CUE points from files
Folks in the bioacoustics space tend to use more ordinary recording devices which can sometimes record important moments as WAVE CUE points.

This changeset adds the ability to read those cue points and also extract  them to a separate file.

Done mainly as an experiment to see if it was useful for someone else.