Extends support for detecting and fixing FL010
Fixes #235

The metadata duration bug fix can now be applied to any firmware. Additionally the fix is now far more intelligent in that it actually counts the total number of samples in the problem file.

Necessary additions include expanded capabilities for the FLAC utility, including the ability to enumerate through frames.