Fixes a big in Flac.CountSamples
Now does a full rescan of frames if the quick tail-only scan returns mixed frames from a flac file. Mixed frames aren't allowed in a stream and it's a good indicator we've hit a desync.

Also improved the micrphone gain and channel parsing a little bit for older firmware log files

Fixes #262