Maintenance
Fixed xunit theory serialization that caused exit code 1 despite all tests passing. This was caused by a legal implicit cast to ReadOnlyMemory<byte> but that type is not supported by xunit as a serialized type.

In the process of working out what went wrong, upgraded dependencies, and .net version, and fixed warnings.