Makes FL011 fix more resiliant
Fixes #346

Mainly: by not throwing in CalculateSampleCountFromFrameList we allow the backup method of counting frames to work. This particular bug was caused because the full file scan method was used to verify FL011, but the Fl010 detection started with the end-only method, and threw when it found invalid frames (WAVE data).

Also fixes a bug where the FL011 fixer in dry mode would touch files that would be made (it left  them empty). No no extra files are produced.