Speculative file locking fix
On intensive runs it seems that sometimes files are still locked for access whenever we attempt to write to them. I've searched extensively for undisposed file streams but as far as I can tell we dispose everything before write access is opened.

This patch seems to reduce the number of file lock errors, but testing is sporadic and unreliable. Speculative.